add reload of command queue after file upload
If a new cmdq.txt was uploaded via the webinterface, the internal CommandQueue needs to be reloaded in order to reflect the changes. Added a public void to do so, perhaps this is usefull later. If the reload is omitted, the uploaded cmdq.txt file will be overwritten on any new command arriving or at reboot of the esp and the contents will never be evaluated.