Added a true "Status" Binary Sensor for HA
Added a default Home Assistant "Status" / Connectivity status binary sensor. This has been added with entity classes so it's added to the correct parts of HA if we ever get the DEVICE added - Status is "device_class: connectivity" and "entity_category: diagnostic". Home assistant does the rest and sets the icon, changes it on up/down, and the payload_on/payload_off sets the correct state to Connected/Disconnected.