Add Support for DS18B20 Temperature Sensor
Extended project to support a single DS18B20 temperature sensor to detect ambient temperature. This will set the ambient temperature from the temperature sensor each time the the periodicTimerFlag is true (defaults to 60 seconds).

If the sensor is not enabled, the system will not attempt to update the ambient temperature.

A digital pin must be set to use the temperature sensor. When using the Wemos D1 Mini, pin D1 has been verified to work well.