Dark mode and other CSS/HTML/JS changes
Primarily wanted to add dark mode, but also took the opportunity to clean up some of the UX/UI of the controller's pages.

At some point this will need to be converted to react. Due to the nature of the initial build of the HTML dark mode is a bit wonky but the primary goal of not being blinded whilst in a dark hot tub or elsewhere is achieved.

Stylisticly some of the CSS and design queues were updated to be cleaner, but work remains. For example, the buttons on the controller for temp/air temp, etc. need be improved.